fix(ci): SMI-5931 resolve retrieval-eval-gate vitest from npm ci
The retrieval-eval-gate job's eval unit-test step pinned vitest via
`npx --yes -p vitest@4.1.2`, orphaned drift from two Dependabot bumps
that moved root's actual vitest to 4.1.10 without ever touching this
hardcoded pin. The mismatch triggers a known npm Arborist ideal-tree
crash (npm/cli#9787, npm/cli#8261) — reproduced deterministically on
2 consecutive CI runs on PR #2068.

Drop the pin and resolve vitest from the job's own npm ci install
instead — the same version every other CI job already uses, and the
same pattern the adjacent unpinned `npx tsx` step in this job already
follows.

Widened during implementation: package-lock.json carried the same
stale nested vitest@4.1.6 duplicate (vs. both package.json files'
4.1.10) not just under packages/doc-retrieval-mcp as originally
scoped in the plan doc, but also under packages/cli,
packages/mcp-server, and packages/website — four copies, not one.
A bare unqualified npx from any of those four directories would have
silently resolved the wrong version instead of crashing. Confirmed
by reproducing the exact edgesOut Arborist crash locally via
`npm dedupe` against the stale lockfile (previously only
'plausible, not reproduced' per the plan's review). Fixed by
surgically removing all four stale nested vitest/@vitest-mocker
lockfile entries (dedup only — no package.json range changes);
verified via a clean npm ci in an isolated tree that all workspace
members now resolve a single hoisted vitest@4.1.10, and that the
actual fixed command (npx vitest run tests/eval/) passes all 10
eval test files (194 tests) under that version.

Also adds a supply-chain scanner test case for the new unqualified
form and a one-line eval/README.md touch so this PR's own
retrieval-eval-gate run actually exercises the fixed step instead of
being skipped by the eval/** path filter.

Splits scripts/tests/check-supply-chain-pins.test.ts (already 684
lines pre-existing, over the 500-line gate) into that file plus a new
check-supply-chain-pins.run-block-installs.test.ts, moving the whole
Check 4 (SMI-4874, workflow npm/npx install audit) section — the
functional area the new test case belongs to — into the sibling file.
Pre-existing violation surfaced by editing this file; fixed in the
same PR per repo policy rather than deferred.

Refs SMI-5931
